requirements.txt
faiss-cpu
gradio==5.25.2
sentence-transformers
torch
transformers
matplotlib
numpy
scipy
python-dotenv
groq
